drivers: mmc: disable CRC
Short for cyclic redundancy check, a common technique for detecting data transmission errors. Transmitted messages are divided into predetermined lengths that are divided by a fixed divisor. According to the calculation, the remainder number is appended into and sent with the message. When the message is received, the phone recalculates the remainder and compares it to the transmitted remainder. If the numbers do not match, an error is detected.

Disabling software CRCs we improve the performace.